Add get by id

+++ Contact-Register/src/ContactRegister.Application/Interfaces/Repositories/IContactRepository.cs
@@ -5,8 +5,8 @@ namespace ContactRegister.Application.Interfaces.Repositories;
 public interface IContactRepository
 {
     public Task AddContactAsync(Contact contact);
-    public Task<Contact?> GetContactByIdAsync(long id);
+    public Task<Contact?> GetContactByIdAsync(int id);
     public Task<List<Contact>> GetContactsAsync();
     public Task UpdateContactAsync(Contact contact);
-    public Task DeleteContactAsync(long id);
+    public Task DeleteContactAsync(int id);
 }
