+++ Contact-Register/src/ContactRegister.Application/Interfaces/Services/IContactService.cs
@@ -6,8 +6,8 @@ namespace ContactRegister.Application.Interfaces.Services;
 public interface IContactService
 {
     public Task<ErrorOr<Success>> AddContactAsync(ContactDto contact);
-    public Task<ErrorOr<ContactDto?>> GetContactByIdAsync(long id);
+    public Task<ErrorOr<ContactDto?>> GetContactByIdAsync(int id);
     public Task<ErrorOr<List<ContactDto>>> GetContactsAsync();
-    public Task<ErrorOr<Success>> UpdateContactAsync(long id, ContactDto contact);
-    public Task<ErrorOr<Success>> DeleteContactAsync(long id);
+    public Task<ErrorOr<Success>> UpdateContactAsync(int id, ContactDto contact);
+    public Task<ErrorOr<Success>> DeleteContactAsync(int id);
 }
